+++ app/src/main/java/com/forrestguice/suntimeswidget/graph/LightMapView.java
@@ -15,7 +15,7 @@
     You should have received a copy of the GNU General Public License
     along with SuntimesWidget.  If not, see <http://www.gnu.org/licenses/>.
 */
-package com.forrestguice.suntimeswidget;
+package com.forrestguice.suntimeswidget.graph;
 
 import android.content.Context;
 import android.graphics.Bitmap;
