+++ app/src/main/res/layout/layout_dialog_lightmap.xml
@@ -63,7 +63,8 @@
         app:layout_constraintStart_toStartOf="parent" app:layout_constraintEnd_toEndOf="parent" />
 
     <!-- lightmap -->
-    <com.forrestguice.suntimeswidget.LightMapView android:id="@+id/info_time_lightmap"
+    <com.forrestguice.suntimeswidget.graph.LightMapView
+        android:id="@+id/info_time_lightmap"
         android:layout_width="0dp" android:layout_height="12dp"
         android:layout_marginTop="@dimen/dialog_margin"
         app:layout_constraintTop_toBottomOf="@+id/info_time_graph"
